requirements.txt
func_timeout==4.3.5
openai==1.55.2
python-dotenv==1.0.1
regex==2024.9.11
sentence_transformers==3.1.1
sql_metadata==2.15.0
stanfordcorenlp==3.9.1.1
torch==2.4.1+cu118
transformers==4.46.0
